*Idle Area

+++ Assets/Scripts/Datas/ValueObject/BuildData.cs
@@ -7,8 +7,8 @@ namespace Datas.ValueObject
     public struct BuildData
     {
         public string BuildName;
-        public int BuildCost;
-        public int GardenCost;
+        public float BuildCost;
+        public float GardenCost;
         public Material BuildMaterial;
         public Material GardenMaterial;
     }
